interface+libraries+tools: Capture this explicitly in lambda copy capture.
Since C++20 it is deprecated to implicitly capture this with [=] lambda
copy expressions.

See: https://en.cppreference.com/w/cpp/language/lambda#Lambda_capture

This patch replaces these occurances with [this] where only this is
required and [=, this] where more scope is required.

In case only this and another variable is required in the scope I chose
to use that variable explicitly in favor of =.

This fixes the GCC14 warning: implicit capture of ‘this’ via ‘[=]’ is deprecated in C++20 [-Wdeprecated].